requirements.txt
streamlit 
pillow
